add decorator to support `as_json`
